Various tweaks to tc_tracks_forecast
- Add two new parameters to fetch_ecmwf(), target_dir and remote_dir which are
passed to fetch_bufr_ftp. Allows for the user to save the raw downloaded
forecast data and to specify which forecast is downloaded from the FTP.
- Replace a few try ... catch statements with if ... else
- Modify fetch_bufr_ftp to always return a list of open file objects. Previously
it might return a list of strings instead, depending on the input parameters. This
makes the logic a little easier to follow, and means that the files always get
closed.
- In _self_check_variable add new parameter varname. It's only there to improve the
error messages. Also improve the error messages with advice on how unexpected they
are.
- Add new method get_value_from_bufr_array which extracts the non-missing value from
an extracted BUFR array. This doesn't change the logic, but moves some repeated
boilerplate code into a separate method.
- In read_one_bufr_tc create the n_ens = len(ens_no)  variable, which simplifies
subsequent code a bit.